refactor(video): extract Runway helpers
